requirements.txt
apify
python-Wappalyzer
requests
jinja2
beautifulsoup4
